Transaction signing (#71)
* move BinarySerializer and BinaryParser

* add STObject

* current state of txn signing

* current state of sign

* current state of sign

* use String instead of Cow<'a, str> for structs using serde_with_tag macro due to lifetime issues when deserializing

* utilize Cow for models

* utilize Cow for models

* cargo fmt

* add simple test

* add devcontainer

* fix async client tests

* fix github workflow

* fix tests with --all-features

* fix --no-default-features tests with embedded-ws

* try fix colliding embebedded-websocket feature with dep: syntax

* fix ambiguous name error

* Revise Models (#74)

* current state

* initial

* initial

* add to changelog

* impl LedgerObject trait

* remove .DS_Store

* add documentation

* cargo fix

* add Default trait to FlagCollection

* refactoring of models

* fix serde tests

* fix serde tests

* Delete src/models/.DS_Store

* remove dafault_none

* resolve comments

* revise dependencies

* remove old clients

* add new clients

* add codec for std embedded websocket

* move websocket exceptions

* add WebsocketBase

* add Client and AsyncClient traits

* add XRPLWebsocketIO as standardized trait

* adjust clients mod

* add XRPLResponse

* revise the Request trait to have methods to get a resut models common fields

* cargo fmt

* add get_random_id utility function

* add xrpl testnet uri

* revise the tests to work with the changes

* error handling

* improve testing

* implement websocket client request

* implement websocket client request for embedded websocket

* update rand trying to get rid of cargo check: OsRng: rand_core::RngCore is not satisfied error

* fix github workflow build and tests

* fix github workflow build and tests

* run linters

* add json rpc client

* add json rpc for std and tests

* cargo fmt

* add pre-commit test with json-rpc-std

* cargo fmt

* cargo fix

* improve client code

* refactor features

* fix tungstenite example

* fix clients

* fix client tests

* fix tests

* rename features in github workflow

* add transaction auto filling as in #76 (#80)

* current state of txn signing

* current state of sign

* current state of sign

* add simple test

* resolve errors caused by updating branch

* resolve errors caused by updating branch

* add autofill_and_sign

* make tx signing work

* improve sign test

* handle unwraps

* add autofill_and_sign

* add sign and submit

* fix errors caused by solving merge conflicts

* rename SerializedDict and SerializedList to STObject and STArray

* rename SerializedDict and SerializedList to STObject and STArray

* handle unwraps

* fix tests

* fix tests

* fix tests

* add multisigning

* add multisigning

* fix tests

* fix tests

---------

+++ src/core/definitions/mod.rs
@@ -65,7 +65,7 @@ pub struct FieldHeader {
 /// let field_instance: FieldInstance =
 ///     FieldInstance::new(&field_info, "Generic", field_header);
 /// ```
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct FieldInstance {
     pub nth: i16,
     pub is_vl_encoded: bool,
@@ -146,10 +146,10 @@ impl ToBytes for FieldHeader {
 
         if self.type_code < 16 {
             if self.field_code < 16 {
-                let shift = self.type_code << 4 | self.field_code;
+                let shift = (self.type_code << 4 | self.field_code) as u8;
                 header_bytes.extend_from_slice(&shift.to_be_bytes());
             } else {
-                let shift = self.type_code << 4;
+                let shift = (self.type_code << 4) as u8;
 
                 header_bytes.extend_from_slice(&shift.to_be_bytes());
                 header_bytes.extend_from_slice(&self.field_code.to_be_bytes());
